add hook for gauge
