+++ packages/react/components/input/gauge/InputGauge.tsx
@@ -1,10 +1,10 @@
 import clsx from 'clsx'
-import React, { useEffect, useMemo } from 'react'
+import React from 'react'
 
 import { hashClass } from '../../../helpers'
-import { TrilogyColor, getColorStyle } from '../../../objects'
-import { Icon, IconColor, IconName, IconSize } from '../../icon'
+import { Icon, IconName, IconSize } from '../../icon'
 import { IValidationRules } from '../InputProps'
+import { useGauge } from '../hook/useGauge'
 
 interface InputGaugeProps {
   validationRules?: IValidationRules
@@ -23,67 +23,16 @@ interface DataVerifyProps {
 }
 
 const InputGauge = ({ validationRules, styled, inputValue }: InputGaugeProps): JSX.Element => {
-  const [points, setPoints] = React.useState<number>(0)
-  const initStateVerifies = { isVerify: false, color: IconColor.NEUTRAL }
-
-  const [isLengthVerify, setIsLengthVerify] = React.useState(initStateVerifies)
-  const [isSpecialCharsVerify, setIsSpecialCharsVerify] = React.useState(initStateVerifies)
-  const [isNumberVerify, setIsNumberVerify] = React.useState(initStateVerifies)
-  const [isUppercaseVerify, setIsUppercaseVerify] = React.useState(initStateVerifies)
-  const [isLowerercaseVerify, setisLowerercaseVerify] = React.useState(initStateVerifies)
-
-  const nbAllVerifies = useMemo(
-    () => (validationRules && Object.values(validationRules).filter((rule) => rule).length) || 0,
-    [validationRules],
-  )
-
-  const widthGauge = React.useMemo(() => {
-    const calc = Number(((points / nbAllVerifies) * 100).toFixed(0))
-    if (calc <= 50 && calc > 0) return '50%'
-    if (calc <= 99 && calc > 50) return '75%'
-    if (calc === 100) return '100%'
-    return '0%'
-  }, [points, nbAllVerifies])
-
-  const colorGauge = React.useMemo(() => {
-    const calc = Number(((points / nbAllVerifies) * 100).toFixed(0))
-    if (calc <= 50 && calc > 0) return getColorStyle(TrilogyColor.ERROR)
-    if (calc <= 99 && calc > 50) return getColorStyle(TrilogyColor.WARNING)
-    if (calc === 100) return getColorStyle(TrilogyColor.SUCCESS)
-    return getColorStyle(TrilogyColor.NEUTRAL_LIGHT)
-  }, [points, nbAllVerifies])
-
-  const LengthvalidationRulesText = useMemo(() => {
-    switch (true) {
-      case validationRules?.length?.min !== undefined && validationRules?.length?.max !== undefined:
-        return `Entre ${validationRules?.length?.min} et ${validationRules?.length?.max} caractères`
-      case validationRules?.length?.min && !validationRules?.length?.max:
-        return `Minimum ${validationRules?.length?.min} caractères`
-      case validationRules?.length?.max && !validationRules?.length?.min:
-        return `Maximum ${validationRules?.length?.max} caractères`
-    }
-  }, [validationRules?.length?.min, validationRules?.length?.max])
-
-  useEffect(() => {
-    const min = validationRules?.length?.min || 0
-    const max = validationRules?.length?.max || ''
-    const regex = new RegExp(`^.{${min},${max}}$`)
-    const validations = []
-
-    validationRules?.specialChars &&
-      validations.push({ test: /[^\w\*]/.test(inputValue), setState: setIsSpecialCharsVerify })
-    validationRules?.number && validations.push({ test: /[0-9]/.test(inputValue), setState: setIsNumberVerify })
-    validationRules?.uppercase && validations.push({ test: /[A-Z]/.test(inputValue), setState: setIsUppercaseVerify })
-    validationRules?.lowercase && validations.push({ test: /[a-z]/.test(inputValue), setState: setisLowerercaseVerify })
-    validationRules?.length && validations.push({ test: regex.test(inputValue), setState: setIsLengthVerify })
-
-    validations.forEach(({ test, setState }) => {
-      if (test) return setState({ isVerify: true, color: IconColor.SUCCESS })
-      return setState(initStateVerifies)
-    })
-
-    setPoints(validations.filter((item) => item.test).length)
-  }, [inputValue, validationRules])
+  const {
+    widthGauge,
+    colorGauge,
+    LengthvalidationRulesText,
+    isLengthVerify,
+    isLowerercaseVerify,
+    isNumberVerify,
+    isSpecialCharsVerify,
+    isUppercaseVerify,
+  } = useGauge({ validationRules, inputValue })
 
   return (
     <div data-testid='security-gauge' className={hashClass(styled, clsx('security-gauge-container'))}>
@@ -162,7 +111,6 @@ const DataVerify = ({
   styled,
 }: DataVerifyProps): JSX.Element | null => {
   if (!display) return null
-  console.log(color)
 
   return (
     <div {...dataAttribute} className={hashClass(styled, clsx('security', classes))}>
